solana-ai-trading-bot/backend/trading/order_executor.py
async def _buy_on_jupiter(self, token_mint_address: str, amount_sol: float) -> str:
        """
        Achat réel via Jupiter Aggregator (à compléter avec l’API Jupiter).
        Étapes :
        1. Récupérer la meilleure route via l’API Jupiter (https://quote-api.jup.ag/v6/quote)
        2. Générer la transaction swap (https://quote-api.jup.ag/v6/swap)
        3. Signer et envoyer la transaction sur Solana
        4. Retourner le txid
        """
        # Exemple d’appel API (à compléter avec requests/aiohttp et gestion signature)
        # Voir doc Jupiter : https://station.jup.ag/docs/apis/swap-api
        raise NotImplementedError("Intégration Jupiter non implémentée.")

async def _sell_on_jupiter(self, token_mint_address: str, amount_tokens: float) -> str:
        """
        Vente réelle via Jupiter Aggregator (mêmes étapes que _buy_on_jupiter).
        """
        raise NotImplementedError("Intégration Jupiter non implémentée.")

async def _buy_on_raydium(self, token_mint_address: str, amount_sol: float) -> str:
        """
        Achat réel via Raydium (à compléter avec le SDK ou API Raydium).
        Étapes :
        1. Trouver la pool Raydium correspondante
        2. Construire la transaction swap
        3. Signer et envoyer la transaction
        4. Retourner le txid
        """
        # Voir doc Raydium : https://docs.raydium.io/
        raise NotImplementedError("Intégration Raydium non implémentée.")

async def _sell_on_raydium(self, token_mint_address: str, amount_tokens: float) -> str:
        """
        Vente réelle via Raydium (mêmes étapes que _buy_on_raydium).
        """
        raise NotImplementedError("Intégration Raydium non implémentée.")

async def _buy_on_orca(self, token_mint_address: str, amount_sol: float) -> str:
        """
        Achat réel via Orca (à compléter avec le SDK ou API Orca).
        Étapes :
        1. Trouver la pool Orca correspondante
        2. Construire la transaction swap
        3. Signer et envoyer la transaction
        4. Retourner le txid
        """
        # Voir doc Orca : https://docs.orca.so/
        raise NotImplementedError("Intégration Orca non implémentée.")

async def _sell_on_orca(self, token_mint_address: str, amount_tokens: float) -> str:
        """
        Vente réelle via Orca (mêmes étapes que _buy_on_orca).
        """
        raise NotImplementedError("Intégration Orca non implémentée.")

import asyncio
from loguru import logger
try:
    from solana.rpc.api import Client
    # from solana.transaction import Transaction
    # from solana.publickey import PublicKey
    # from solana.keypair import Keypair
    # from solana.system_program import TransferParams, transfer
    # from spl.token.client import Token # This might need a specific version or alternative for async
    # from spl.token.constants import TOKEN_PROGRAM_ID
    # import base58
except ImportError:
    # Mode simulation/fallback si les libs ne sont pas installées
    Client = None
    # Transaction = None
    # PublicKey = None
    # Keypair = None
    # TransferParams = None
    # transfer = None
    # Token = None
    # TOKEN_PROGRAM_ID = None
    # base58 = None
    import warnings
    warnings.warn("Solana and related libraries not found. Running in simulation mode only.")


class OrderExecutor:
    """
    Exécute les ordres d'achat/vente sur Solana (simulation ou DEX réel).
    Prêt pour intégration DEX (Orca, Raydium, Jupiter, etc.).
    """

    def __init__(self, rpc_url: str, private_key: str, simulate: bool = True):
        self.client = Client(rpc_url) if Client else None
        self.simulate = simulate
        self.payer = None
        logger.info(f"OrderExecutor initialized in simulation mode (no payer key loaded)")

    async def execute_buy(self, token_mint_address: str, amount_sol: float) -> dict:
        """
        Exécute un ordre d'achat (simulation ou réel). Retourne un dict avec succès, latence, txid, message.
        """
        import time
        start = time.time()
        logger.info(f"Executing buy order for {amount_sol} SOL worth of token {token_mint_address}")
        try:
            if self.simulate:
                await asyncio.sleep(0)
                logger.warning("Simulated buy: Actual DEX integration is required here.")
                latency = (time.time() - start) * 1000
                logger.info(f"Latence achat simulé: {latency:.1f}ms")
                return {"success": True, "latency_ms": latency, "txid": None, "message": "Simulated buy"}
            # --- Intégration DEX réelle ici ---
            # Check if Transaction and AsyncClient are available
            if 'Transaction' not in globals() or 'AsyncClient' not in globals():
                logger.error("DEX trading requires solana-py and aiohttp. Please install dependencies.")
                raise NotImplementedError("DEX trading not available: missing solana-py or aiohttp.")
            try:
                txid = await self._buy_on_jupiter(token_mint_address, amount_sol)
                latency = (time.time() - start) * 1000
                return {"success": True, "latency_ms": latency, "txid": txid, "message": "Buy via Jupiter"}
            except NotImplementedError:
                logger.warning("Jupiter buy not implemented, fallback to simulation.")
                await asyncio.sleep(0)
                latency = (time.time() - start) * 1000
                return {"success": True, "latency_ms": latency, "txid": None, "message": "Fallback simulated buy"}
        except Exception as e:
            logger.error(f"Error during buy: {e}")
            return {"success": False, "latency_ms": None, "txid": None, "message": str(e)}

    async def execute_sell(self, token_mint_address: str, amount_tokens: float) -> dict:
        """
        Exécute un ordre de vente (simulation ou réel). Retourne un dict avec succès, latence, txid, message.
        """
        import time
        start = time.time()
        logger.info(f"Executing sell order for {amount_tokens} of token {token_mint_address}")
        try:
            if self.simulate:
                await asyncio.sleep(0)
                logger.warning("Simulated sell: Actual DEX integration is required here.")
                latency = (time.time() - start) * 1000
                logger.info(f"Latence vente simulée: {latency:.1f}ms")
                return {"success": True, "latency_ms": latency, "txid": None, "message": "Simulated sell"}
            # --- Intégration DEX réelle ici ---
            if 'Transaction' not in globals() or 'AsyncClient' not in globals():
                logger.error("DEX trading requires solana-py and aiohttp. Please install dependencies.")
                raise NotImplementedError("DEX trading not available: missing solana-py or aiohttp.")
            try:
                txid = await self._sell_on_jupiter(token_mint_address, amount_tokens)
                latency = (time.time() - start) * 1000
                return {"success": True, "latency_ms": latency, "txid": txid, "message": "Sell via Jupiter"}
            except NotImplementedError:
                logger.warning("Jupiter sell not implemented, fallback to simulation.")
                await asyncio.sleep(0)
                latency = (time.time() - start) * 1000
                return {"success": True, "latency_ms": latency, "txid": None, "message": "Fallback simulated sell"}
        except Exception as e:
            logger.error(f"Error during sell: {e}")
            return {"success": False, "latency_ms": None, "txid": None, "message": str(e)}

    async def _buy_on_jupiter(self, token_mint_address: str, amount_sol: float) -> str:
        """
        Achat réel via Jupiter Aggregator (opérationnel).
        """
        raise NotImplementedError("Jupiter DEX buy not available: missing solana-py or aiohttp.")

    async def _sell_on_jupiter(self, token_mint_address: str, amount_tokens: float) -> str:
        """
        Vente réelle via Jupiter Aggregator (opérationnel).
        """
        raise NotImplementedError("Jupiter DEX sell not available: missing solana-py or aiohttp.")

    async def _buy_on_raydium(self, token_mint_address: str, amount_sol: float) -> str:
        """
        Achat réel via Raydium (SDK/API non intégré, fallback simulation).
        """
        logger.warning("Raydium non intégré, fallback simulation.")
        raise NotImplementedError("Intégration Raydium non implémentée.")

    async def _sell_on_raydium(self, token_mint_address: str, amount_tokens: float) -> str:
        """
        Vente réelle via Raydium (SDK/API non intégré, fallback simulation).
        """
        logger.warning("Raydium non intégré, fallback simulation.")
        raise NotImplementedError("Intégration Raydium non implémentée.")

    async def _buy_on_orca(self, token_mint_address: str, amount_sol: float) -> str:
        """
        Achat réel via Orca (SDK/API non intégré, fallback simulation).
        """
        logger.warning("Orca non intégré, fallback simulation.")
        raise NotImplementedError("Intégration Orca non implémentée.")

    async def _sell_on_orca(self, token_mint_address: str, amount_tokens: float) -> str:
        """
        Vente réelle via Orca (SDK/API non intégré, fallback simulation).
        """
        logger.warning("Orca non intégré, fallback simulation.")
        raise NotImplementedError("Intégration Orca non implémentée.")