solana-ai-trading-bot/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
websockets==12.0
sqlalchemy==2.0.30
asyncio==3.4.3
python-jose[cryptography]==3.3.0
pycryptodome==3.20.0
pysolana==0.1.1
base58==2.1.1
PyNaCl==1.5.0
loguru==0.7.2
python-multipart==0.0.9
networkx